separate `useDeleteMediaMutation` and `withDeleteMedia`
